+++ packages/xrpl/src/utils/stringConversion.ts
@@ -1,19 +1,17 @@
-import {
-  bytesToHex,
-  hexToBytes,
-  TextDecoder,
-  TextEncoder,
-} from '@xrplf/isomorphic/utils'
+import { stringToHex, hexToString } from '@xrplf/isomorphic/utils'
 
 /**
  * Converts a string to its hex equivalent. Useful for Memos.
  *
  * @param string - The string to convert to Hex.
  * @returns The Hex equivalent of the string.
+ *
+ * @deprecated use `@xrplf/isomorphic/utils`'s `stringToHex`
+ *
  * @category Utilities
  */
 function convertStringToHex(string: string): string {
-  return bytesToHex(new TextEncoder().encode(string))
+  return stringToHex(string)
 }
 
 /**
@@ -22,10 +20,13 @@ function convertStringToHex(string: string): string {
  * @param hex - The hex to convert to a string.
  * @param encoding - The encoding to use. Defaults to 'utf8' (UTF-8). 'ascii' is also allowed.
  * @returns The converted string.
+ *
+ * @deprecated use `@xrplf/isomorphic/utils`'s `hexToString`
+ *
  * @category Utilities
  */
 function convertHexToString(hex: string, encoding = 'utf8'): string {
-  return new TextDecoder(encoding).decode(hexToBytes(hex))
+  return hexToString(hex, encoding)
 }
 
 export { convertHexToString, convertStringToHex }